+++ src/Beef.Core/Caching/Policy/AbsoluteCachePolicy.cs
@@ -51,7 +51,7 @@ namespace Beef.Caching.Policy
         {
             get
             {
-                DateTime now = DateTime.Now;
+                DateTime now = Entities.Cleaner.Clean(DateTime.Now);
                 if (Expiry != null && now <= Expiry)
                     return false;
 
@@ -91,7 +91,7 @@ namespace Beef.Caching.Policy
         /// </summary>
         public void Reset()
         {
-            DateTime now = DateTime.Now;
+            DateTime now = Entities.Cleaner.Clean(DateTime.Now);
             DateTime? expiry = now.Add(_duration);
 
             if (_randomizerOffset.HasValue)