+++ src/Beef.Grpc/CHANGELOG.md
@@ -2,5 +2,8 @@
 
 Represents the **NuGet** versions.
 
+## v3.1.2
+- *Enhancement:* Updated all dependent NuGet packages to their latest respective version.
+
 ## v3.1.1
 - *New:* Initial publish to GitHub. New capability to support gRPC.